[feat] AuthService 클래스 수정 및 통일 #64
- ErrorStatus 추가
