+++ src/main/java/chungbazi/chungbazi_be/global/apiPayload/code/status/ErrorStatus.java
@@ -31,6 +31,9 @@ public enum ErrorStatus implements BaseErrorCode {
     INVALID_VALUE(HttpStatus.BAD_REQUEST, "USER400", "유효하지 않은 입력값입니다."),
     INVALID_NICKNAME(HttpStatus.BAD_REQUEST, "USER400", "이미 존재하는 닉네임입니다."),
     INVALID_CHARACTER(HttpStatus.BAD_REQUEST, "CHARACTER400", "유효하지 않은 캐릭터입니다."),
+    ALREADY_EXISTS_EMAIL(HttpStatus.BAD_REQUEST,"USER4005","중복된 이메일 입니다."),
+    INVALID_CREDENTIALS(HttpStatus.BAD_REQUEST,"USER4006","잘못된 비밀번호 입니다."),
+    PASSWORD_MISMATCH(HttpStatus.BAD_REQUEST,"USER4007","비밀번호와 확인 비밀번호가 일치하지 않습니다."),
 
     //Policy
     CATEGORY_CODE_NOT_FOUND(HttpStatus.BAD_REQUEST, "POLICY5001", "존재하지 않는 정책 코드 입니다."),
